# packages/api-client/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "GOD_MODE_CODE API",
    "description" : "Backend for GOD_MODE_CODE.",
    "version" : "0.0.1"
  },
  "servers" : [ {
    "url" : "/",
    "description" : "Same origin as the application"
  } ],
...
  "paths" : {
...
    "/api/patterns/{slug}/challenges" : {
      "post" : {
        "tags" : [ "Patterns" ],
        "summary" : "Be handed this Pattern to solve",
        "description" : "Records an Issue against the User — who got what, when, and until when — and returns the Pattern with its Scaffold and Example Tests.\n\nUnlike a Passage, the player names which one. There is nothing to shop for: a Pattern is a technique to practise, not a score to farm.\n\nAsking again abandons whatever Challenge the User was holding, in either Discipline, so nobody can hold several at once and submit whichever went best (ADR-0003). The window is flat rather than scaled by length: a Pattern is answered by thinking, and there is nothing about it to time by the character.\n",
        "operationId" : "request",
        "parameters" : [ {
          "name" : "gmc_recognition",
          "in" : "cookie",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "The Challenge, and when it stops being answerable",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SolveChallenge"
                }
              }
            }
          },
          "401" : {
            "description" : "This browser is nobody yet — create a User first"
          },
          "404" : {
            "description" : "No activated Pattern has that slug"
          }
        }
      }
    },
...
    "/api/challenges" : {
      "post" : {
        "tags" : [ "Challenges" ],
        "summary" : "Be handed a Passage to transcribe",
        "description" : "Records an Issue against the User — who got what, when, and until when — and returns the Passage. Which Passage arrives is the server's choice.\n\nAsking again abandons the Challenge the User was holding, so a player can skip a Passage they do not fancy, and nobody can hold several at once and submit whichever went best (ADR-0003).\n",
        "operationId" : "request_1",
        "parameters" : [ {
          "name" : "gmc_recognition",
          "in" : "cookie",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ChallengeRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "The Challenge, and when it stops being answerable",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Challenge"
                }
              }
            }
          },
          "401" : {
            "description" : "This browser is nobody yet — create a User first"
          },
          "404" : {
            "description" : "That Discipline has no Passages. Code never will (ADR-0004)"
          },
          "429" : {
            "description" : "Too many Challenges have been asked for by this User or from this address recently"
          }
        }
      },
      "delete" : {
        "tags" : [ "Challenges" ],
        "summary" : "Give up the Challenge you are holding",
        "description" : "Voids whatever Challenge this User is holding, without recording a Run against it. Leaving mid-Run (the Escape key) calls this so the Issue stops being live immediately, rather than sitting live until the next Challenge is requested.\n\nAnswers 204 whether or not anything was actually live to give up, so a player who has already finished, or never started, is not told anything went wrong.\n",
        "operationId" : "abandon",
        "parameters" : [ {
          "name" : "gmc_recognition",
          "in" : "cookie",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Nothing is live for this User anymore"
          },
          "401" : {
            "description" : "This browser is nobody yet"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ExampleTest" : {
        "type" : "object",
        "description" : "A test case shown to the player before they start",
        "properties" : {
          "name" : {
            "type" : "string",
            "description" : "What the case is about",
            "example" : "the pair is the first two numbers"
          },
          "call" : {
            "type" : "string",
            "description" : "The call the submitted source must satisfy",
            "example" : "pair_sum([2, 7, 11, 15], 9)"
          },
          "expected" : {
            "type" : "string",
            "description" : "The answer the call has to produce",
            "example" : [ 0, 1 ]
          }
        },
        "required" : [ "call", "expected", "name" ]
      },
      "Pattern" : {
        "type" : "object",
        "description" : "A Pattern to solve, as the player is shown it",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "slug" : {
            "type" : "string",
            "description" : "Stable identifier, also what the judge knows this Pattern by",
            "example" : "hash-map-seen-lookup"
          },
          "name" : {
            "type" : "string",
            "example" : "Store what you've seen, look up what you need"
          },
          "family" : {
            "type" : "string",
            "description" : "The grouping a Pattern belongs to",
            "enum" : [ "HASH_MAP", "TWO_POINTERS", "SLIDING_WINDOW", "STACK", "HEAP", "BINARY_SEARCH", "GRAPH", "DYNAMIC_PROGRAMMING" ]
          },
          "seniority" : {
            "type" : "string",
            "description" : "The difficulty band of a Challenge",
            "enum" : [ "JUNIOR", "SENIOR", "PRINCIPAL" ]
          },
          "prompt" : {
            "type" : "string",
            "description" : "The technique and what to do with it, in the player's words"
          },
          "scaffold" : {
            "type" : "string",
            "description" : "The read-only lines above the editable region",
            "example" : "def pair_sum(numbers, target):"
          },
          "exampleTests" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ExampleTest"
            }
          }
        },
        "required" : [ "exampleTests", "family", "id", "name", "prompt", "scaffold", "seniority", "slug" ]
      },
      "SolveChallenge" : {
        "type" : "object",
        "description" : "A Pattern to solve, and the Issue that recorded it going out",
        "properties" : {
          "issueId" : {
            "type" : "string",
            "format" : "uuid",
            "description" : "What the client hands back when the Solve Run is submitted"
          },
          "pattern" : {
            "$ref" : "#/components/schemas/Pattern"
          },
          "expiresAt" : {
            "type" : "string",
            "format" : "date-time",
            "description" : "After this moment the Challenge can no longer be answered"
          }
        },
        "required" : [ "expiresAt", "issueId", "pattern" ]
      },
...
      "ChallengeRequest" : {
        "type" : "object",
        "description" : "A request for a Challenge in one Discipline",
        "properties" : {
          "discipline" : {
            "type" : "string",
            "description" : "One of the three ways to play",
            "enum" : [ "QUOTES", "PROSE", "CODE" ]
          }
        },
        "required" : [ "discipline" ]
      },
      "Challenge" : {
        "type" : "object",
        "description" : "A Passage to transcribe, and the Issue that recorded it going out",
        "properties" : {
          "issueId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "passage" : {
            "$ref" : "#/components/schemas/Passage"
          },
          "expiresAt" : {
            "type" : "string",
            "format" : "date-time",
            "description" : "After this moment the Challenge can no longer be answered"
          }
        },
        "required" : [ "expiresAt", "issueId", "passage" ]
      },
      "Passage" : {
        "type" : "object",
        "description" : "A fixed piece of text to be transcribed",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "discipline" : {
            "type" : "string",
            "description" : "One of the three ways to play",
            "enum" : [ "QUOTES", "PROSE", "CODE" ]
          },
          "text" : {
            "type" : "string",
            "description" : "The text to transcribe, printable ASCII throughout",
            "example" : "Call me Ishmael."
          },
          "attribution" : {
            "type" : "string",
            "description" : "Who said or wrote it",
            "example" : "Herman Melville, Moby-Dick, 1851"
          },
          "characterCount" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "How many characters the text is",
            "example" : 302
          }
        },
        "required" : [ "attribution", "characterCount", "discipline", "id", "text" ]
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "description" : "The identity provider's ID token, validated by the backend as a resource server (ADR-0011).",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}